Adjusts outputs to improve usability
TL;DR
-----

Outputs a unique ID and the Kubernetes secret storing the key

Details
-------

Adds outputs for the Kubernetes certificate containing the key
and for a unique identifier. The identifier at this point is
just the private key, though it could probably be parsed out of
it more specifically in the future.

To Do
-----

Simplify the ID