No longer creates a Kubernetes secret
TL;DR
-----

Shifts responsibility for K8s secrets to the caller

Details
-------

Updates the module so it doesn't need to interact with a Kubernetes
cluster to create secrets. This both makes it more readily reusable
and more directly allows it to be used prior to a Kubernetes dialtone
in the lab environment.